// src/app/models/productDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductDTO { 
    id?: number;
    name?: string;
    categoryId?: number;
    categoryName?: string;
    description?: string;
    price?: number;
    imageUrl?: string;
    stockQuantity?: number;
    ingredients?: string;
}