// src/app/models/categoryDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CategoryDTO { 
    id?: number;
    name?: string;
    description?: string;
    imageUrl?: string;
}